TAB is bound to sp-up-sexp and evil-jump-forward (which was standard)
I had <tab> bound to sp-up-exp at first but this did not allow me to
tab-complete in the mini-buffer. After reverting this change it works again.
Only problem I still have is that when I TAB at the closing brace/parenthesis I
get 'jumped' towards the end of my file.