Remove ranger folder
the ranger folder has been removed from the config because
the standard config of ranger changes quite a lot through updates.
It is easier to just get the latest config and work from there.

This can be done as follows:
```
ranger --copy-config=all
```
this copies all the ranger config into the `.config/ranger` folder.